Add charset to Content-Type when possible
This change causes a "charset" parameter to be appended to the Content-Type header set when serving files from the cache, if possible.
Currently, only text files (text/* MIME type) are recognized and automatically marked as UTF-8 (this behaviour depends on the "mime" npm package).
Each cache entry now includes a "charset" property, either undefined or a string.