Switch to normal transfer encoding and send Content-Length for gzipped files
This change allows clients (such as XHR requests) to obtain the total size of the served file, e.g. to display a progress bar.
One caveat that applies is that the HTTP spec does not define if Content-Length refers to the compressed or uncompressed content.
Here, we arbitrarily pick the compressed content (body length verbatim), as it is the most readily-available measure for the client.